+++ pyanalyze/test_name_check_visitor.py
@@ -24,6 +24,7 @@ from .value import (
     AsyncTaskIncompleteValue,
     CallableValue,
     DictIncompleteValue,
+    KVPair,
     KnownValue,
     MultiValuedValue,
     NewTypeValue,
@@ -117,6 +118,7 @@ def _make_module(code_str: str) -> types.ModuleType:
         AsyncTaskIncompleteValue=AsyncTaskIncompleteValue,
         CallableValue=CallableValue,
         DictIncompleteValue=DictIncompleteValue,
+        KVPair=KVPair,
         GenericValue=GenericValue,
         KnownValue=KnownValue,
         MultiValuedValue=MultiValuedValue,
@@ -524,12 +526,18 @@ def run():
             s = {a, b}
             assert_is_value(s, SequenceIncompleteValue(set, [UNANNOTATED, UNANNOTATED]))
             z = {a: b}
-            assert_is_value(z, DictIncompleteValue(dict, [(UNANNOTATED, UNANNOTATED)]))
+            assert_is_value(
+                z, DictIncompleteValue(dict, [KVPair(UNANNOTATED, UNANNOTATED)])
+            )
             q = {a: 3, b: 4}
             assert_is_value(
                 q,
                 DictIncompleteValue(
-                    dict, [(UNANNOTATED, KnownValue(3)), (UNANNOTATED, KnownValue(4))]
+                    dict,
+                    [
+                        KVPair(UNANNOTATED, KnownValue(3)),
+                        KVPair(UNANNOTATED, KnownValue(4)),
+                    ],
                 ),
             )
 
@@ -932,7 +940,8 @@ class TestUnwrapYield(TestNameCheckVisitorBase):
 
             vals3 = yield {1: square.asynq(1)}
             assert_is_value(
-                vals3, DictIncompleteValue(dict, [(KnownValue(1), TypedValue(int))])
+                vals3,
+                DictIncompleteValue(dict, [KVPair(KnownValue(1), TypedValue(int))]),
             )
 
             vals4 = yield {i: square.asynq(i) for i in ints}
@@ -1378,9 +1387,9 @@ class TestIterationTarget(TestNameCheckVisitorBase):
                 DictIncompleteValue(
                     dict,
                     [
-                        (KnownValue(1), KnownValue(1)),
-                        (KnownValue(2), KnownValue(2)),
-                        (KnownValue(3), KnownValue(3)),
+                        KVPair(KnownValue(1), KnownValue(1)),
+                        KVPair(KnownValue(2), KnownValue(2)),
+                        KVPair(KnownValue(3), KnownValue(3)),
                     ],
                 ),
             )
@@ -1391,12 +1400,7 @@ class TestIterationTarget(TestNameCheckVisitorBase):
             lst = []
             if cond:
                 lst.append("x")
-            assert_is_value(
-                lst,
-                MultiValuedValue(
-                    [SequenceIncompleteValue(list, [KnownValue("x")]), KnownValue([])]
-                ),
-            )
+            assert_is_value(lst, KnownValue(["x"]) | KnownValue([]))
             for c in lst:
                 assert_is_value(c, KnownValue("x"))
 
@@ -2165,10 +2169,62 @@ class TestAugAssign(TestNameCheckVisitorBase):
 class TestUnpacking(TestNameCheckVisitorBase):
     @assert_passes()
     def test_dict_unpacking(self):
-        def capybara():
+        from typing import Dict
+        from typing_extensions import TypedDict, NotRequired
+
+        class FullTD(TypedDict):
+            a: int
+            b: str
+
+        class PartialTD(TypedDict):
+            a: int
+            b: NotRequired[str]
+
+        def capybara(d: Dict[str, int], ftd: FullTD, ptd: PartialTD):
             d1 = {1: 2}
             d2 = {3: 4, **d1}
-            assert_is_value(d2, TypedValue(dict))
+            assert_is_value(
+                d2,
+                DictIncompleteValue(
+                    dict,
+                    [
+                        KVPair(KnownValue(3), KnownValue(4)),
+                        KVPair(KnownValue(1), KnownValue(2)),
+                    ],
+                ),
+            )
+            assert_is_value(
+                {1: 2, **d},
+                DictIncompleteValue(
+                    dict,
+                    [
+                        KVPair(KnownValue(1), KnownValue(2)),
+                        KVPair(TypedValue(str), TypedValue(int), is_many=True),
+                    ],
+                ),
+            )
+            assert_is_value(
+                {1: 2, **ftd},
+                DictIncompleteValue(
+                    dict,
+                    [
+                        KVPair(KnownValue(1), KnownValue(2)),
+                        KVPair(KnownValue("a"), TypedValue(int)),
+                        KVPair(KnownValue("b"), TypedValue(str)),
+                    ],
+                ),
+            )
+            assert_is_value(
+                {1: 2, **ptd},
+                DictIncompleteValue(
+                    dict,
+                    [
+                        KVPair(KnownValue(1), KnownValue(2)),
+                        KVPair(KnownValue("a"), TypedValue(int)),
+                        KVPair(KnownValue("b"), TypedValue(str), is_required=False),
+                    ],
+                ),
+            )
 
     @assert_passes()
     def test_iterable_unpacking(self):